refactor(maps): extract MonsterBatchDialog from ActiveMapStage
Move the self-contained 1–20 batch-placement modal into its own file. No
behavior change; testids preserved (verified by the monster-tokens e2e).
